New refactored screens and functionalities

+++ app/src/main/java/com/snowtouch/groupmarket/core/domain/model/Advertisement.kt
@@ -9,6 +9,6 @@ data class Advertisement(
     val title: String? = null,
     val images: List<String>? = null,
     val description: String? = null,
-    val price: Int? = null,
+    val price: Double? = null,
     val postDate: Date? = null
 )
